--- a/src/Microsoft.AspNet.Server.Kestrel/Http/ListenerContext.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/Http/ListenerContext.cs
@@ -11,13 +11,15 @@ namespace Microsoft.AspNet.Server.Kestrel.Http
     {
         public ListenerContext()
         {
-            Memory2 = new MemoryPool2();
+            InputMemory = new MemoryPool2();
+            OutputMemory = new MemoryPool2();
         }
 
         public ListenerContext(ServiceContext serviceContext) 
             : base(serviceContext)
         {
-            Memory2 = new MemoryPool2();
+            InputMemory = new MemoryPool2();
+            OutputMemory = new MemoryPool2();
         }
 
         public ListenerContext(ListenerContext listenerContext)
@@ -26,7 +28,8 @@ namespace Microsoft.AspNet.Server.Kestrel.Http
             ServerAddress = listenerContext.ServerAddress;
             Thread = listenerContext.Thread;
             Application = listenerContext.Application;
-            Memory2 = listenerContext.Memory2;
+            InputMemory = listenerContext.InputMemory;
+            OutputMemory = listenerContext.OutputMemory;
             Log = listenerContext.Log;
         }
 
@@ -36,6 +39,8 @@ namespace Microsoft.AspNet.Server.Kestrel.Http
 
         public Func<Frame, Task> Application { get; set; }
 
-        public MemoryPool2 Memory2 { get; set; }
+        public MemoryPool2 InputMemory { get; set; }
+
+        public MemoryPool2 OutputMemory { get; set; }
     }
 }