… src/Microsoft.AspNet.Server.Kestrel/Infrastructure/KestrelThread.cs
… src/Microsoft.AspNet.Server.Kestrel/KestrelEngine.cs
… src/Microsoft.AspNet.Server.Kestrel/Networking/UvWriteReq.cs
… test/Microsoft.AspNet.Server.KestrelTests/MultipleLoopTests.cs
… test/Microsoft.AspNet.Server.KestrelTests/NetworkingTests.cs
… test/Microsoft.AspNet.Server.KestrelTests/SocketOutputTests.cs