Has a bug and code is messy but @ 80% cpu

~/wrk$ wrk -c 1024 -d 30 -t 32 ...url.. -s ./scripts/pipeline.lua -- 16
Running 30s test @ http://url
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    20.30ms   27.86ms 508.17ms   95.38%
    Req/Sec    30.19k     6.65k   53.32k    77.83%
  27798979 requests in 30.10s, 3.42GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: **923596.22**
Transfer/sec:    116.27MB

**Has a bug** that occasionally happens (neeeed sleep)

>    Microsoft.AspNet.Server.Kestrel.FunctionalTests.ResponseTests.LargeDownload [FAIL]
     System.Net.Http.HttpRequestException : An error occurred while sending the request.
     ---- System.Net.Http.WinHttpException : The server returned an invalid or unrecognized response