@benaadams 2015-10-29
Now

~/wrk$ wrk -c 1024 -d 30 -t 32 http://url -s ./scripts/pipeline.lua -- 16
Running 30s test @ http://url
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.15ms   15.10ms 278.97ms   81.29%
    Req/Sec    31.72k     5.98k   66.75k    72.80%
  29491574 requests in 30.10s, 3.63GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: **979796.86**
Transfer/sec:    123.34MB