@benaadams 2015-10-29
Using this benchmark change https://github.com/aspnet/benchmarks/pull/23