@benaadams 2015-10-29
Needs 2 runs to warm up; watching network, probably needs some more memory clean up

~/wrk$ wrk -c 1024 -d 30 -t 32 http://url -s ./scripts/pipeline.lua -- 16
Running 30s test @ http://url
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.14ms   23.11ms 312.42ms   90.21%
    Req/Sec    31.32k     8.61k   87.35k    80.25%
  29846046 requests in 30.10s, 3.67GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: **991579.43**
Transfer/sec:    124.83MB