@AspNetSmurfLab 2015-10-29
@benaadams what were the results on your rig before this change? I've tried it here and it seems to substantially change the CPU profile, such that I now need to use many more connections (1024) than before in order to get high numbers. I'm now maxing CPU and hitting 505k RPS, which is about 7% more than before, although before we weren't quite maxing CPU.