@benaadams 2015-10-29
@AspNetSmurfLab very weird will look into it a bit more; before on 16 core 2GHz VM (Azure G4)

`wrk -c 256 -d 30 -t 32 ... -s ./scripts/pipeline.lua -- 16` 616K RPS 75% CPU
`wrk -c 1024 -d 30 -t 32 ... -s ./scripts/pipeline.lua -- 16` 587K RPS 73% CPU

So decreasing scaling; after (with the async local removal tweak in benchmarking)

`wrk -c 256 -d 30 -t 32 ... -s ./scripts/pipeline.lua -- 16` 523K RPS 60% CPU
`wrk -c 1024 -d 30 -t 32 ... -s ./scripts/pipeline.lua -- 16` 941K RPS 91% CPU

The decrease at 256 seems to be latency; so the before latency is 4.64ms and after it is 14.6ms; also this version needs a couple `wrk` runs to warm up; possibly to populate the `MemoryPool2`.

Will dig a little deeper, do a clean build etc and see if I can improve the latency.