@benaadams 2015-10-29
Also which arch? Am testing coreclr x64