@benaadams 2015-10-30
Closing for now as working on some code that will be ugly and don't want it as an open PR